[SMA-230] Refactor structure related to user/service
